refactor(tls): move net.py into strix/utils
